# doc/api.yaml
...
openapi: 3.0.1
info:
  description: |
    These are API for the app WASAPhoto, the application for the
    2023/24 WASA Course.
  title: WASAPhoto
  version: 1.0.2

# Global application of the bearer authorization
security:
  - Authorization: []

# Server
servers:
  - url: http://localhost:3000
    description: Local server

# Components definition
components:
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      description: "Bearer token for authorization"
  #-------------------------- Definition of the schemas --------------------------
  schemas:
...
    Timestamp:
      description: |
        The timestamp of the object
      type: string
      format: date-time
      example: "2025-01-01 00:00:00"

    #-------------User related-------------
    UserId:
      description: |
            Identifies a single user.
      type: integer
      minimum: 1
      example: 15
      readOnly: true
...
    PostId:
      description: "Identifies a single post"
      type: integer
      minimum: 1
      example: 17
      readOnly: true
    PostDescription:
      description: "The description of a post"
      type: string
      example: "Me at the zoo"
      minLength: 0
      maxLength: 500
      pattern: "^[a-zA-Z0-9\\s\\S]*$"
    PostPhoto:
      description: "The photo of the post"
      type: string
      format: binary
      minLength: 0
      maxLength: 10000000

    #-------------Comment related-------------
    CommentId:
      description: "Identifies a single comment"
      type: integer
      minimum: 1
      example: 17
      readOnly: true
    CommentBody:
      description: "The body of the comment"
      type: string
      example: "LOL! Really liked it"
      minLength: 0
      maxLength: 255
      pattern: "^[a-zA-Z0-9\\s\\S]*$"

    #--------------------------Objects--------------------------#
    # User object
...
    Post:
      title: "Post"
      description: "This object represents a single post by a user"
      type: object
      properties:
        id: {$ref: "#/components/schemas/PostId"}
        user_id: {$ref: "#/components/schemas/UserId"}
        description: {$ref: "#/components/schemas/PostDescription"}
        photo: {$ref: "#/components/schemas/PostPhoto"}
        timestamp: {$ref: "#/components/schemas/Timestamp"}

    # Comment object
    Comment:
      title: "Comment"
      description: "This object represents a single comment"
      type: object
      properties:
        id: {$ref: "#/components/schemas/CommentId"}
        userid:  {$ref: "#/components/schemas/UserId"}
        post_owner: {$ref: "#/components/schemas/UserId"}
        post_id: {$ref: "#/components/schemas/PostId"}
        content: {$ref: "#/components/schemas/CommentBody"}
        timestamp: {$ref: "#/components/schemas/Timestamp"}

  #-------------------------- Definition of the parameters --------------------------#
  parameters:
    user_id:
      schema: {$ref: "#/components/schemas/UserId"}
      name: "user_id"
      in: path
      required: true
      description: Id of a single user, used to identify a specific user.
    post_id:
      schema: {$ref: "#/components/schemas/PostId"}
      name: "post_id"
      in: path
      required: true
      description: Id of a single post, used to identify a specific post.
    comment_id:
      schema: {$ref: "#/components/schemas/CommentId"}
      name: "comment_id"
      in: path
      required: true
      description: Id of a single comment, used to identify a specific comment.
  #-------------------------- Definition of the responses --------------------------#
  responses:
    NotFound:
      description: Resource was not found
      content:
        text/plain:
          schema:
            type: string
            description: The error
            minLength: 0
            maxLength: 512
            pattern: "^[a-zA-Z0-9\\s\\S]*$"
paths:
...
  /users/{user_id}/photos/{post_id}:
    parameters:
        - {$ref: "#/components/parameters/post_id"}
        - {$ref: "#/components/parameters/user_id"}
    get:
      tags: ["post"]
      operationId: getPhoto
      summary: Get a specific photo
      description: |
        Return a specific post of a user.
      responses:
        "200":
          description: Return post.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Post"}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      tags: ["post"]
      operationId: deletePhoto
      summary: Delete a photo
      description: |
        Delete the photo with the specified ID.
        Doesn't return anything
      responses:
        "200":
          description: Photo deleted.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: Photo with id 294829 was deleted
        "404": {$ref: "#/components/responses/NotFound"}
...
  /users/{user_id}/photos/{post_id}/likes/{user_like_id}:
    parameters:
        - {$ref: "#/components/parameters/post_id"}
        - {$ref: "#/components/parameters/user_id"}
        - schema: {$ref: "#/components/schemas/UserId"}
          name: "user_like_id"
          in: path
          required: true
          description: Id of the user that likes/unlikes the photo.
    put:
      tags: ["post like"]
      operationId: likePhoto
      summary: Like a photo
      description: The user with the "user_like_id" ID likes the photo with "post_id" ID.
      responses:
        "201":
          description: Photo liked.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: Post with id 9284929 was liked
        "403":
          description: Photo already liked.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: Post with id 9284929 is already liked
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      tags: ["post like"]
      operationId: unlikePhoto
      summary: Unlike a photo
      description: The user with the "user_like_id" ID unlikes the photo with "post_id" ID.
      responses:
        "200":
          description: Photo unliked.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: Post with id 9284929 was unliked
        "400":
          description: Photo already unliked.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: Post with id 9284929 is already unliked
        "404": {$ref: "#/components/responses/NotFound"}
  /users/{user_id}/photos/{post_id}/comments:
    parameters:
        - {$ref: "#/components/parameters/post_id"}
        - {$ref: "#/components/parameters/user_id"}
    get:
      tags: ["post comment"]
      operationId: getComments
      summary: Get comments of a photo
      description: |
        Return the list of comments of the photo with "post_id" ID.
      responses:
        "200":
          description: Return comments.
          content:
            application/json:
              schema:
                type: object
                description: Array of comments of the post
                properties:
                  comments:
                    type: array
                    items: { $ref: "#/components/schemas/Comment"}
                    description: Array of comments
                    minItems: 0
                    maxItems: 999999
        "404": {$ref: "#/components/responses/NotFound"}

    post:
      operationId: commentPhoto
      tags: ["post comment"]
      summary: Post comment
      description: |
        The user who calls this comments the photo with "post_id" ID.
        The comment must be in the request body.
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Comment"}

      responses:
        "201":
          description: Photo commented.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: Post with id 2402029 was commented
        "404": {$ref: "#/components/responses/NotFound"}
  /users/{user_id}/photos/{post_id}/comments/{comment_id}:
    parameters:
        - {$ref: "#/components/parameters/post_id"}
        - {$ref: "#/components/parameters/user_id"}
        - {$ref: "#/components/parameters/comment_id"}
    delete:
      tags: ["post comment"]
      operationId: deleteComment
      summary: Delete comment
      description: |
        The user who calls this delete his comment with "comment_id" ID.
        The comment must be his.
      responses:
        "200":
          description: Comment deleted.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: The comment with id 389282 was deleted
        "401":
          description: Unauthorized.
          content:
            text/plain:
              schema:
                type: string
                description: Response string
                example: The comment with id 299482 is not yours
        "404": {$ref: "#/components/responses/NotFound"}